+++ jdklog-api/src/main/java/org/jdkstack/jdklog/logging/api/metainfo/Record.java
@@ -1,6 +1,5 @@
 package org.jdkstack.jdklog.logging.api.metainfo;
 
-import java.time.Instant;
 import java.util.List;
 import java.util.Map;
